pyproject.toml
[project]
name = "kvyo-mcp"
version = "0.1.0"
description = "MCP server for klaviyo dev stuff"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.10.1",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "llama-index>=0.10.0",
]
